+++ ruby/lib/gitlab/rails_logger.rb
@@ -1,0 +1,10 @@
+require 'logger'
+
+module Rails
+  LOGGER = Logger.new(STDOUT)
+  LOGGER.level = 'info'
+
+  def self.logger
+    LOGGER
+  end
+end